+++ packages/auditable-item-graph-models/src/models/IAuditableItemGraphPatchOperation.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import type { IPatchOperation } from "@gtsc/core";
+import type { IPatchOperation } from "@twin.org/core";
 
 /**
  * The patch operation for JSON diffs.